re-haul modules and day2
start to make re-usable code.
in this case I am trying to re-use the same way to read the input files
import them as raw data, and then process them as it needs be by using modules.

solution to day2 also posted, tests to be done as well as improvements to it.